refactor: move concurrency-cap semaphore from openai_utils to openai_model
Per review: keep NeMoGymAsyncOpenAI as a thin aiohttp transport wrapper
and put the optional max_concurrent_requests semaphore on the model
server instead. Other model servers (vllm, azure) don't need this knob
today, and non-model callers of NeMoGymAsyncOpenAI (tavily, browsecomp)
hit different upstreams — so the cap belongs at the policy layer, not
the transport layer.

- nemo_gym/openai_utils.py: drop max_concurrent_requests field,
  _semaphore, _get_semaphore, the ConfigDict(arbitrary_types_allowed),
  and the per-method `async with semaphore or nullcontext()` wrapping.
- responses_api_models/openai_model/app.py: build the semaphore in
  model_post_init from config.max_concurrent_requests; wrap
  client.create_chat_completion / create_response with it via _bound().
- Move the two semaphore tests from tests/unit_tests/test_openai_utils.py
  to responses_api_models/openai_model/tests/test_app.py.

Retry-policy changes stay in openai_utils (where they belong).
