+++ tests/unit_tests/test_openai_utils.py
@@ -12,8 +12,6 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-import asyncio
-
 import pytest
 from pydantic import ValidationError
 
@@ -23,37 +21,6 @@ from nemo_gym.openai_utils import NeMoGymAsyncOpenAI, NeMoGymResponseCreateParam
 class TestOpenAIUtils:
     async def test_NeMoGymAsyncOpenAI(self) -> None:
         NeMoGymAsyncOpenAI(api_key="abc", base_url="https://api.openai.com/v1")
-
-    def test_semaphore_disabled_by_default(self) -> None:
-        client = NeMoGymAsyncOpenAI(api_key="abc", base_url="https://api.openai.com/v1")
-        assert client.max_concurrent_requests is None
-        assert client._get_semaphore() is None
-
-    @pytest.mark.asyncio
-    async def test_semaphore_caps_concurrency(self) -> None:
-        client = NeMoGymAsyncOpenAI(
-            api_key="abc",
-            base_url="https://api.openai.com/v1",
-            max_concurrent_requests=2,
-        )
-        sem = client._get_semaphore()
-        assert sem is not None
-        # Same instance returned across calls (lazy init only fires once).
-        assert client._get_semaphore() is sem
-
-        in_flight = 0
-        peak = 0
-
-        async def worker() -> None:
-            nonlocal in_flight, peak
-            async with sem:
-                in_flight += 1
-                peak = max(peak, in_flight)
-                await asyncio.sleep(0.01)
-                in_flight -= 1
-
-        await asyncio.gather(*(worker() for _ in range(8)))
-        assert peak == 2
 
 
 class TestNeMoGymResponseCreateParamsNonStreaming:
